cleaning and hash fix

+++ modules/Cat/Cat.cpp
@@ -9,7 +9,7 @@ using namespace std;
 
 // Compute hash of moduleName at compile time, so the moduleName string don't show in the binary
 constexpr std::string_view moduleName = "cat";
-constexpr unsigned long moduleHash = djb2(moduleName);
+constexpr unsigned long long moduleHash = djb2(moduleName);
 
 #ifdef _WIN32
