+++ modules/Evasion/Evasion.cpp
@@ -17,7 +17,7 @@
 using namespace std;
 
 constexpr std::string_view moduleName = "evasion";
-constexpr unsigned long moduleHash = djb2(moduleName);
+constexpr unsigned long long moduleHash = djb2(moduleName);
 
 
 #ifdef _WIN32
