+++ modules/Inject/Inject.cpp
@@ -9,7 +9,7 @@
 using namespace std;
 
 constexpr std::string_view moduleName = "inject";
-constexpr unsigned long moduleHash = djb2(moduleName);
+constexpr unsigned long long moduleHash = djb2(moduleName);
 
 #ifdef _WIN32
