+++ modules/Run/Run.cpp
@@ -17,7 +17,7 @@
 using namespace std;
 
 constexpr std::string_view moduleName = "run";
-constexpr unsigned long moduleHash = djb2(moduleName);
+constexpr unsigned long long moduleHash = djb2(moduleName);
 
 #define BUFSIZE 4096
